--HG--
branch : spec_SFEperl-mouse

+++ specs/SFEperl-data-dumper.spec
@@ -1,20 +1,21 @@
 %include Solaris.inc
-%include default-depend.inc
 
 %define build584 0
+%define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 0
 
 %define cpan_name Data-Dumper
 %define sfe_cpan_name data-dumper
+%define ips_cpan_name data-dumper
 
 Summary:               stringified perl data structures, suitable for both printing and eval
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
-Version:               2.154
-IPS_component_version: 2.154
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
+Version:               2.161
+IPS_component_version: 2.161
 License:               unknown
 URL:                   https://metacpan.org/pod/Data::Dumper
 Source0:               http://cpan.metacpan.org/authors/id/S/SM/SMUELLER/Data-Dumper-%{version}.tar.gz
@@ -25,25 +26,41 @@ stringified perl data structures, suitable for both printing and eval
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          stringified perl data structures, suitable for both printing and eval
 BuildRequires:    runtime/perl-584 = *
-#BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/test-simple-584
 Requires:         runtime/perl-584 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 
 %description 584
 stringified perl data structures, suitable for both printing and eval
 %endif
 
+%if %{build510}
+%package 510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
+Summary:          stringified perl data structures, suitable for both printing and eval
+BuildRequires:    runtime/perl-510 = *
+BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/test-simple-510
+Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+
+%description 510
+stringified perl data structures, suitable for both printing and eval
+%endif
+
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          stringified perl data structures, suitable for both printing and eval
 BuildRequires:    runtime/perl-512 = *
-# BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/test-simple-512
 Requires:         runtime/perl-512 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 
 %description 512
 stringified perl data structures, suitable for both printing and eval
@@ -51,38 +68,42 @@ stringified perl data structures, suitable for both printing and eval
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          stringified perl data structures, suitable for both printing and eval
 BuildRequires:    runtime/perl-516 = *
-# BuildRequires:    library/perl-5/extutils-makemaker-516
+BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 
 %description 516
 stringified perl data structures, suitable for both printing and eval
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          stringified perl data structures, suitable for both printing and eval
-BuildRequires:    runtime/perl-520 = *
-#Buildrequires:    library/perl-5/extutils-makemaker-520
-Buildrequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-simple-522
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 
-%description 520
+%description 522
 stringified perl data structures, suitable for both printing and eval
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
     perl_ver=$1
+    test=$2
     bindir="/usr/perl5/${perl_ver}/bin"
     vendor_dir="/usr/perl5/vendor_perl/${perl_ver}"
 
@@ -90,13 +111,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test
     make pure_install
 }
 
 build_with_build.pl_for() {
     perl_ver=$1
+    test=$2
     bindir="/usr/perl5/${perl_ver}/bin"
     vendor_dir="/usr/perl5/vendor_perl/${perl_ver}"
 
@@ -105,34 +130,70 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
-  if [ -f Makefile.PL ];
-  then
-    build_with_makefile.pl_for $*
-  elif [ -f Build.PL ];
+  if [ -f Build.PL ];
   then
     build_with_build.pl_for $*
+  elif [ -f Makefile.PL ];
+  then
+    build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
-
+# To build without test, pass 'without_test' to build_for commaond.
+# like 'build_for version without_test'
 %if %{build584}
 build_for 5.8.4
+%endif
+
+%if %{build510}
+build_for 5.10
 %endif
 
 %if %{build512}
@@ -143,8 +204,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -163,7 +224,7 @@ rm -rf %{buildroot}
 
 %files
 %defattr(0755,root,bin,-)
-# %{_datadir}/man/man3perl
+# %{_datadir}/man
 
 %if %{build584}
 %files 584
@@ -172,6 +233,16 @@ rm -rf %{buildroot}
 /usr/perl5/vendor_perl/5.8.4
 %if %{include_executable}
 /usr/perl5/5.8.4
+%endif
+%endif
+
+%if %{build510}
+%files 510
+%defattr(0755,root,bin,-)
+%dir %attr (0755, root, sys) /usr
+/usr/perl5/vendor_perl/5.10
+%if %{include_executable}
+/usr/perl5/5.1.0
 %endif
 %endif
 
@@ -195,17 +266,18 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 2.161, package for perl-522 is added and for perl-520 is obsolete
 * Tue Nov 03 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - initial commit
