+++ specs/SFEperl-test-exception.spec
@@ -4,7 +4,8 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name Test-Exception
@@ -14,8 +15,8 @@
 Summary:               Test exception-based code
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               0.40
-IPS_component_version: 0.40
+Version:               0.43
+IPS_component_version: 0.43
 License:               perl_5
 URL:                   https://metacpan.org/pod/Test::Exception
 Source0:               http://cpan.metacpan.org/authors/id/E/EX/EXODIST/Test-Exception-%{version}.tar.gz
@@ -31,6 +32,13 @@ Summary:          Test exception-based code
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/test-simple-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-584
+BuildRequires:    library/perl-5/exporter-584
+BuildRequires:    library/perl-5/sub-uplevel-584
+BuildRequires:    library/perl-5/test-harness-584
+BuildRequires:    library/perl-5/test-simple-584
+%endif
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-584
@@ -49,6 +57,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Test exception-based code
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/carp-510
+BuildRequires:    library/perl-5/exporter-510
+BuildRequires:    library/perl-5/sub-uplevel-510
+BuildRequires:    library/perl-5/test-harness-510
 BuildRequires:    library/perl-5/test-simple-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
@@ -69,6 +82,13 @@ Summary:          Test exception-based code
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/test-simple-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-512
+BuildRequires:    library/perl-5/exporter-512
+BuildRequires:    library/perl-5/sub-uplevel-512
+BuildRequires:    library/perl-5/test-harness-512
+BuildRequires:    library/perl-5/test-simple-512
+%endif
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-512
@@ -88,6 +108,14 @@ Summary:          Test exception-based code
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-516
+BuildRequires:    library/perl-5/exporter-516
+BuildRequires:    library/perl-5/sub-uplevel-516
+BuildRequires:    library/perl-5/test-harness-516
+BuildRequires:    library/perl-5/test-simple-516
+%endif
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-516
@@ -100,29 +128,36 @@ Requires:         library/perl-5/test-simple-516
 Test exception-based code
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Test exception-based code
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-simple-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-522
+BuildRequires:    library/perl-5/exporter-522
+BuildRequires:    library/perl-5/sub-uplevel-522
+BuildRequires:    library/perl-5/test-harness-522
+BuildRequires:    library/perl-5/test-simple-522
+%endif
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/carp-520
-Requires:         library/perl-5/exporter-520
-Requires:         library/perl-5/sub-uplevel-520
-Requires:         library/perl-5/test-harness-520
-Requires:         library/perl-5/test-simple-520
+Requires:         library/perl-5/carp-522
+Requires:         library/perl-5/exporter-522
+Requires:         library/perl-5/sub-uplevel-522
+Requires:         library/perl-5/test-harness-522
+Requires:         library/perl-5/test-simple-522
 
-%description 520
+%description 522
 Test exception-based code
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -135,8 +170,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -151,7 +190,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -168,7 +207,7 @@ modify_bin_dir() {
     then
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -186,7 +225,11 @@ modify_man_dir() {
             mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
             rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
         fi
-        rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
     fi
 }
 
@@ -221,8 +264,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -283,18 +326,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu May 25 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 0.43
 * Fri Nov 13 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 0.40 and build packages for perl-510, perl-516 and perl-520
 * Tue Jun 19 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
