+++ specs/SFEperl-test-pod-coverage.spec
@@ -4,15 +4,17 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name Test-Pod-Coverage
 %define sfe_cpan_name test-pod-coverage
+%define ips_cpan_name test-pod-coverage
 
 Summary:               Check for pod coverage in your distribution
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
 Version:               1.10
 IPS_component_version: 1.10
 License:               artistic_2
@@ -25,12 +27,17 @@ Check for pod coverage in your distribution
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Check for pod coverage in your distribution
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/test-simple-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/pod-coverage-584
+BuildRequires:    library/perl-5/test-simple-584
+%endif
 Requires:         runtime/perl-584 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-coverage-584
 Requires:         library/perl-5/test-simple-584
 
@@ -40,12 +47,15 @@ Check for pod coverage in your distribution
 
 %if %{build510}
 %package 510
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Check for pod coverage in your distribution
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
 BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/pod-coverage-510
+BuildRequires:    library/perl-5/test-simple-510
 Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-coverage-510
 Requires:         library/perl-5/test-simple-510
 
@@ -55,12 +65,17 @@ Check for pod coverage in your distribution
 
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Check for pod coverage in your distribution
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/test-simple-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/pod-coverage-512
+BuildRequires:    library/perl-5/test-simple-512
+%endif
 Requires:         runtime/perl-512 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-coverage-512
 Requires:         library/perl-5/test-simple-512
 
@@ -70,12 +85,18 @@ Check for pod coverage in your distribution
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Check for pod coverage in your distribution
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/pod-coverage-516
+BuildRequires:    library/perl-5/test-simple-516
+%endif
 Requires:         runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pod-coverage-516
 Requires:         library/perl-5/test-simple-516
 
@@ -83,25 +104,30 @@ Requires:         library/perl-5/test-simple-516
 Check for pod coverage in your distribution
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Check for pod coverage in your distribution
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/pod-coverage-520
-Requires:         library/perl-5/test-simple-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-simple-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/pod-coverage-522
+BuildRequires:    library/perl-5/test-simple-522
+%endif
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/pod-coverage-522
+Requires:         library/perl-5/test-simple-522
 
-%description 520
+%description 522
 Check for pod coverage in your distribution
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -114,8 +140,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} = 'without_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -130,17 +160,47 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} = 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
@@ -152,7 +212,8 @@ build_for() {
     build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
@@ -173,14 +234,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -235,18 +296,21 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Apr 27 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- fix build
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Wed Nov 11 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 1.10 and build packages for perl-510, perl-516 and perl-520
 * Thu Jun 14 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
